Tilemap Things
Fixed the tiles to be even and made them readable when designing the level. Final version of tiles will not have the arrows, they are just for us and not the player.

IMPORTANT: To work on the game you MUST open the "Scenes" folder in the project window and open the scene "Level1." Otherwise, it'll look like a BLANK PROJECT.